third_party/WebKit/LayoutTests/svg/custom/svg-parse-overflow-2-expected.txt

CONSOLE ERROR: line 9: Error: <rect> attribute height: Expected length, "-105000000000000…".
